# backend/src/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: SparkRentals REST API
  version: 1.0.0
  description: |
    SparkRentals scooter-sharing platform API.

    ## Authentication
    All endpoints require an `api_key` query parameter (or in body).
    User endpoints additionally require `x-access-token` JWT header.
    Admin endpoints require admin JWT.

    ## Response Format
    Single item: `{ data: {...}, meta: { requestId } }`
    List: `{ data: [...], meta: { requestId, page: { limit, offset, total, hasMore } } }`
    Error: `{ error: { code, message, details, requestId } }`

    ## Rate Limits
    - Global: 1000 req/min
    - Auth endpoints: 10 attempts/15min
    - OTP sending: 1/min
    - Registrations: 5/hour
    - Mutations: 60/min
servers:
  - url: http://localhost:8393/v1
    description: Local development
  - url: https://api.sparkrentals.uz/v1
    description: Production

tags:
  - name: auth
    description: Authentication (login, register, password reset)
  - name: trips
    description: Trip lifecycle (reserve, start, end, cancel)
  - name: scooters
    description: Scooter management
  - name: users
    description: User accounts
  - name: transactions
    description: Payments and transaction history
  - name: support
    description: Support tickets
  - name: notifications
    description: Push notifications and device management
  - name: promocodes
    description: Promo codes and referral program
  - name: discovery
    description: Find nearby scooters, QR scan
  - name: geofencing
    description: Zone detection and speed limits
  - name: iot
    description: IoT endpoints for scooter firmware (ESP32)
  - name: juicers
    description: Scooter charging team
  - name: mechanics
    description: Maintenance team and parts inventory
  - name: stats
    description: Admin analytics (admin only)
  - name: views
    description: View API (BFF) — aggregated endpoints for screens
  - name: system
    description: System info and health

paths:
  /health:
    get:
      tags: [system]
      summary: Health check
      security: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: ok }
                  timestamp: { type: string, format: date-time }
                  version: { type: string }

  /auth/login/server/user:
    post:
      tags: [auth]
      summary: Login user with email + password
      security: [{ apiKey: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      type: { type: string, example: success }
                      message: { type: string }
                      token: { type: string }
                      user:
                        type: object
                        properties:
                          id: { type: string }
                          email: { type: string }
        '401':
          $ref: '#/components/responses/Unauthorized'

  /auth/phone/send-code:
    post:
      tags: [auth]
      summary: Send SMS OTP code to phone
      description: |
        Sends a 6-digit code via SMS. Rate limited: 1 per minute per IP.
        In dev mode (SMS_PROVIDER=console), the code is returned in response.
      security: [{ apiKey: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone]
              properties:
                phone: { type: string, example: "+998901234567" }
                purpose: { type: string, enum: [login, password_reset, phone_verify], default: login }
      responses:
        '200':
          description: Code sent
        '429':
          $ref: '#/components/responses/RateLimited'

  /auth/phone/login:
    post:
      tags: [auth]
      summary: Login or register with phone + SMS code
      description: If phone is new, auto-registers a user.
      security: [{ apiKey: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone, code]
              properties:
                phone: { type: string }
                code: { type: string }
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      is_new_user: { type: boolean }
                      access_token: { type: string }
                      refresh_token: { type: string }
                      expires_in: { type: integer, example: 900 }
        '401':
          $ref: '#/components/responses/Unauthorized'

  /auth/refresh:
    post:
      tags: [auth]
      summary: Refresh access token
      security: [{ apiKey: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        '200':
          description: New tokens
        '401':
          $ref: '#/components/responses/Unauthorized'

  /trips:reserve:
    post:
      tags: [trips]
      summary: Reserve a scooter
      description: |
        Reserves a scooter for 10 minutes. The scooter's status changes to 'reserved'.
        If the user already has an active or reserved trip, returns 409.
      security: [{ apiKey: [], userToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [scooter_id]
              properties:
                scooter_id: { type: string }
      responses:
        '201':
          description: Reserved
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      trip_id: { type: string }
                      expires_at: { type: string, format: date-time }
                      expires_in_seconds: { type: integer, example: 600 }
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

  /trips:start:
    post:
      tags: [trips]
      summary: Start an active trip from reservation
      security: [{ apiKey: [], userToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [trip_id]
              properties:
                trip_id: { type: string }
      responses:
        '200': { description: Trip started }
        '410': { description: Reservation expired }

  /trips:end:
    post:
      tags: [trips]
      summary: End active trip with cost calculation
      description: |
        Calculates cost based on city rates and end zone:
          cost = max(0, fixedRate + time × minutes - discount + fee)

        - End in parking zone: discount
        - End in bonus zone: bigger discount
        - End in no-parking zone: fee
        - End outside any zone: fee (treated as street)
      security: [{ apiKey: [], userToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [trip_id]
              properties:
                trip_id: { type: string }
                end_coordinates:
                  type: object
                  properties:
                    longitude: { type: string }
                    latitude: { type: string }
                photo_url: { type: string }
      responses:
        '200':
          description: Trip ended
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      trip_id: { type: string }
                      duration_min: { type: integer }
                      cost: { type: number }
                      end_zone: { type: string }
                      end_battery: { type: number }
                      new_balance: { type: number }

  /trips/history:
    get:
      tags: [trips]
      summary: Get user's trip history (paginated)
      security: [{ apiKey: [], userToken: [] }]
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 25, maximum: 100 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
        - name: sort
          in: query
          schema: { type: string, default: "-created_at" }
      responses:
        '200':
          description: Trip list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListResponse'

  /views/dashboard:
    get:
      tags: [views]
      summary: Admin dashboard aggregated view
      description: |
        Returns multiple sections in one request (BFF pattern).
        Sections: main (B0), revenue (B1), fleet (B1).
        Each section has independent cache TTL.
      security: [{ apiKey: [], adminToken: [] }]
      parameters:
        - name: sections
          in: query
          schema: { type: string, example: "main,revenue,fleet" }
          description: Comma-separated list of sections to return
      responses:
        '200':
          description: Dashboard data
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      main: { type: object }
                      revenue: { type: object }
                      fleet: { type: object }
                  meta:
                    type: object
                    properties:
                      sections:
                        type: object
                        additionalProperties:
                          type: object
                          properties:
                            ms: { type: integer }
                            cache: { type: string, enum: [HIT, MISS, ERROR] }
                            error: { type: string, nullable: true }
                      totalMs: { type: integer }
                      partial: { type: boolean }
                      requestId: { type: string }

  /views/trip-detail/{tripId}:
    get:
      tags: [views]
      summary: Trip detail aggregated view
      description: Returns trip + user + scooter + city in one request.
      security: [{ apiKey: [], userToken: [] }]
      parameters:
        - name: tripId
          in: path
          required: true
          schema: { type: string }
        - name: sections
          in: query
          schema: { type: string, example: "main,scooter,user,city" }
      responses:
        '200':
          description: Trip detail with related entities
        '404':
          $ref: '#/components/responses/NotFound'

  /scooters:
    get:
      tags: [scooters]
      summary: List scooters (admin)
      security: [{ apiKey: [], adminToken: [] }]
      parameters:
        - { name: status, in: query, schema: { type: string } }
        - { name: owner, in: query, schema: { type: string }, description: city_id }
        - { name: limit, in: query, schema: { type: integer, default: 50 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses:
        '200':
          description: Scooter list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListResponse'

  /discovery/nearest:
    get:
      tags: [discovery]
      summary: Find nearest available scooters
      security: [{ apiKey: [], userToken: [] }]
      parameters:
        - { name: lat, in: query, required: true, schema: { type: number } }
        - { name: lng, in: query, required: true, schema: { type: number } }
        - { name: radius_km, in: query, schema: { type: number, default: 2 } }
        - { name: limit, in: query, schema: { type: integer, default: 10, maximum: 50 } }
      responses:
        '200':
          description: Nearest scooters sorted by distance

  /iot/telemetry:
    post:
      tags: [iot]
      summary: Scooter telemetry push (from ESP32)
      description: |
        Called by scooter firmware to push GPS + battery + speed updates.
        Authenticates by scooter MAC + per-scooter secret.
      security: [{ apiKey: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [scooter_mac, secret]
              properties:
                scooter_mac: { type: string }
                secret: { type: string }
                coordinates:
                  type: object
                  properties:
                    longitude: { type: string }
                    latitude: { type: string }
                battery: { type: number }
                speed: { type: number }
                status: { type: string }
      responses:
        '200': { description: Telemetry received }
        '401': { description: Invalid secret }
        '404': { description: Scooter not provisioned }

components:
  securitySchemes:
    apiKey:
      type: apiKey
      in: query
      name: api_key
    userToken:
      type: apiKey
      in: header
      name: x-access-token
    adminToken:
      type: apiKey
      in: header
      name: x-access-token
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: TRIP_NOT_FOUND }
            message: { type: string }
            details: { type: object }
            requestId: { type: string }
    ListResponse:
      type: object
      properties:
        data:
          type: array
          items: { type: object }
        meta:
          type: object
          properties:
            requestId: { type: string }
            page:
              type: object
              properties:
                limit: { type: integer }
                offset: { type: integer }
                total: { type: integer }
                hasMore: { type: boolean }
                nextOffset: { type: integer, nullable: true }
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: Conflict
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    RateLimited:
      description: Too many requests
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }